refactor: use same convention for facet list component
